requirements.txt
Jinja2>=3.0.3
marko>=1.1.0
python-dateutil>=2.8.2
pytz>=2021.3
pyScss>=1.3.7
PyYAML>=5.4.1
requests>=2.26.0
